SPUI-1010: fix all subtask in 1010

+++ lib/app/screens/portfolio/helper/max_currency_apy.dart
@@ -3,8 +3,8 @@ import '../../../shared/models/currency_model.dart';
 double maxCurrencyApy(List<CurrencyModel> currencies) {
   var maxApy = 0.0;
   for (final element in currencies) {
-    if (element.apy > maxApy) {
-      maxApy = element.apy;
+    if (element.apy.toDouble() > maxApy) {
+      maxApy = element.apy.toDouble();
     }
   }
   return maxApy;
