+++ lib/app/shared/features/wallet/view/components/wallet_body/empty_earn_wallet_body.dart
@@ -1,3 +1,4 @@
+import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
@@ -17,7 +18,7 @@ class EmptyEarnWalletBody extends HookWidget {
   }) : super(key: key);
 
   final String assetName;
-  final double apy;
+  final Decimal apy;
 
   @override
   Widget build(BuildContext context) {
